requirements.txt
pandas==2.2.1
numpy==1.26.4
requests==2.31.0
python-dotenv==1.0.1
matplotlib==3.8.3 